requirements.txt
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
pyyaml>=6.0,<7.0
rich>=13.0,<14.0
litellm>=1.0,<2.0
pytest>=8.0,<9.0
